Add Cart Counter Icon and Update Home AppBar
Created a new `CartCounterIcon` widget in `lib/common/widgets/product/cart/cart_counter.dart` to display a shopping bag icon with a counter badge. This widget is used to indicate the number of items in the cart.

Refactored the `Home` screen in `lib/features/shop/screens/home/home.dart` to use a new `THomeAppBar` widget located in `lib/features/shop/screens/home/widgets/home_appbar.dart`. The new `THomeAppBar` widget includes the cart counter icon and updates the title display.

This change improves the visual consistency and functionality of the app bar in the home screen.